fix/perf: try to prevent bufferbloat lag in channels
Make tracking channels bounded and use try_send, so frames will be dropped instead of buffered during a lag event.

Switch patteval_call and patteval_update to be bounded (but still use blocking send), so that any lag events will be propagated to the caller. Not sure if this is good or bad, but it could prevent massive growth of the channels during lag events?